Update README with a generic workaround
It is possible to use an output that is not plugged in, like a spare VGA, HDMI or DP port. This servers as a good enough workaround for non Intel devices that do not support virtual outputs.
Also add some information on AMD virtual displays and why they are not a solution in this case.